Add some end-of-game code (untested)
In sc_counts() view, for games with fixed final years, provide the
right number of formsets for the remaining years.
Add DrawProposal.save() to set Game.is_finished if the draw vote passed.
In CentreCount.clean(), check that the year isn't beyond the fixed
final year for the round.
Add CentreCount.save() to set Game.is_finished if the final year has
been completed or somebody won the game.