Add round scoring
Add a force_recalculate parameter to Game.scores() and change the
algorithm to retrieve the scores from the database unless the
game is not yet finished or force_recalculate is True.
In Game.save(), if the game is finished, recalculate the scores and
update them in the database.
Add Round.scores() and Round.save() with ismilar behaviour to the above.
Correct RScoringSystem.scores() now that we have code that calls it.
In the tournament standings template, right align all the table data.
Change tournament_scores() to call the new Round.scores() method
rather than retrieving scores from the database. Also truncate all
scores to 2 decimal places.