Support different versions of the boardgame
Different versions use different colours for the great powers.
It would be confusing to display an SC chart using the Avalon Hill
colour scheme next to an image of the game being played on the
Gibson's board, for example.

Create two new classes - GameSet and SetPower. GameSet represents
a version of the board game. SetPower associates a GreatPower with a
GameSet, and assigns a colour to that power in that set. Then we add
a the_set attribute to the Game class to specify which version of the
game each game is being played with.

The rest is tweaks to the admin to allow the addition of new GameSets,
to the SC chart code to pick up the new colours, to the game creation
form to allow a GameSet to be specified, to the initial data (we add
two GameSets), and to the tests.

I left the original colour attribute on GreatPower to be removed in
a later commit, just in case.