Views: Delete objects that fail validation
If we call get_or_create() and then full_clean() and the latter raises
ValidationError, we need to call delete() to remove the created object.
Otherwise it continues to exist (presumably only in memory and not
in the database, but that's still enough for it to appear in queries, etc).