Fix game updates
In the view, change create_games() to change the player of any
existing GamePlayer for this game and power rather than creating a
new one.
In the model, change GamePlayer.clean() to ensure that we don't
raise ValidationError due to an overlap between one instance and
itself. Also make the "overlap" error message much more verbose
and fix a couple of typos in other error messages.

It's now possible to update games using the form for creating games.