Models: Add support for unpublished tournaments
Add Tournament.is_published, defaulted to False. While this is False,
a Tournament is only visible to logged-in users.

The remainder of the changes are to URLs and views to enforce that:
Scrap TourneyIndexView and TourneyDetailView.
Add tournament_index() which now constructs two lists, one of all published
tournaments and a second of all unpublished tournaments visible to the
current user.
Add get_visible_tournament_or_404() to retrieve the specified Tournament
and to check whether it should be visible to the current user. If not,
raise Http404. Call this from all views into a Tournament.
Update the URLs to reflect the new views.
Update the tournament index template to display the two lists now
passed to it.