DrawProposal.clean(): More tests and resulting bugfixes
Add a check for a successful DrawProposal before the last SC
count for the game.
Add tests to cover all the ways DrawProposal.clean() can fail.
Specify proposers when creating DrawProposals in tests.
Fix powers set so that it contains the actual powers rather than
their pks - this ensure that later tests for powers in the set
behave correctly.
Get SC counts that were valid for the year of the DrawProposal
rather than the latest year of the game (this allows DrawProposals
to be created retroactively if needed).