Models: Add methods to find player positions
Tournament.positions(), Game.positions() and TournamentPlayer.position().
These are analagous to Tournament.scores() and Game.scores(), except
that rather than the absolute scores they return the positions of
the players or powers (respectively). Where scores are equal, they are
given the same position.
TournamentPlayer.position() is simply the position of that specific
player in the tournament (if it ended now).