Views: Re-order game SC chart screen
New ordering is as follows:
First, a paragraph with the game result or "Game Ongoing".
Then, the table with a column per player.
Within the table, first row is now the score, and the remaining
rows are SC counts in reverse years order, so the most recent is
at the top.

This should fix Bug #21, because the most important information
is always at the top and therefore always visible.