Models: Fix add_local_player_bg()
This was broken in two ways: First, because the file split meant
that it didn't have access to constants like WIN and LOSS and second
because it assumed that all 7 powers survived if there was no draw
or solo.

Added Game.survivors() to easily find out who survived a game.
Added TO_GAME_RESULT() to convert integers to DRAW_n constants.