Views: Avoid AttributeError in roll call form
As the comment says, current_round() could return None, so it's a
bad idea to call number() on that...
Note that the old comment said that we get a read-only form. That
doesn't seem to be the case (I tried setting the round number to both
-1 and 999, and saw no difference). I suspect this means that something
else is broken here. For now, I just removed the comment.