Add Game.create_or_update_sc_counts_from_ownerships()
This method either updates existing CentreCounts for the specified
Game and year to match those detailed in corresponding
SupplyCentreOwnerships or it creates them.
It raises an exception if no SupplyCentreOwnerships are found. This
could probably be even more restrictive (should never be fewer than the
home SC total), but this will do for now - it's much more likely that
somebody mis-types a year than that they only create a few of those
objects.
Also alter the SC ownership form to call the new method after adding
or updating all the SupplyCentreOwnerships for a game year. This should
mean that you can either just enter SC counts or just enter SC ownerships
and both methods should work fine (in the latter case, the system will
enter the SC counts automatically).