Templates: Hide useless links
In the game detail page, only display the link to the SC ownership
page if we have post-1900 SC ownership info. Similarly, don't display
the link to the latest position if we only have S1901.
In the tournament view page, add a note that the SC ownership info
may not be populated, but keep the link always present.