Models: Don't add local player background
This was a good thing when it was difficult to get data added to the WDD.
Unfortunately, it's difficult to identify tournaments that are present in
both (in this case, the lcoal one is "Cascadia Open" and the WDD version
is "Cascadia Open 2017", so we ended up with two copies of the
PlayerTournamentRanking in some cases. As it's now easier to get data
into the WDD, just pulling from there should be fine.