Models: Include players who haven't yet attended a round
Tournament.scores() (and hence Tournament.positions_and_scores()) was
treating TournamentPlayers without corresponding RoundPlayers differently
if the tournament was ongoing or ended.
When it has ended, we just retrieve each TournamentPlayer.score, which
defaults to zero. So people who didn't attend any rounds are included with
a score of zero.
When the tournament is ongoing, though (or hasn't started), we were
calculating the scores using the RoundPlayers and therefore omitting
people who hadn't (yet) attended any rounds (everyone if the tournament
hasn't started yet).
So explictly add those people in with scores of zero.
This then simplifies tournament_scores(), because KeyError can no longer
happen when indexing into the dict returned by positions_and_scores()
(that code was wrong anyway).