Cleanup after running pyflakes
Pyflakes reports a number of issues, the majority of which are
just stylistic. This commit fixes almost all of them.
Mostly these replace "import *" with named imports and remove things that
weren't used.
There are three actual bugfixes included:
- validate_vote_count() was raising a non-existent exception
- there were two different test_wdd_background_awards_invalid() methods
- roll_call() was using an undefined variable to add errors to the form
Unfixed are two "imported but unused" in models.py, which are both
now commented (they're used in one of the migrations, and I'm a little
wary of touching those generated files).