Code cleanup
Mostly as suggested by pylint3.
Lots of whitespace and docstrings, plus a few coding convention things.
The two main classes in background were renamed to remove the spaces
from the names.
Some wrapping of long lines.
Missing docstrings
tests of len(sequence) against zero
no need for an "else" if the "if" ends with a return