Views: Add a game overview cycle, similar to tournament overview
In game_sc_owners(), if there is no ownership info for the current
game year and we're refreshing, do an immediate redirect.
Similarly, in game)image(), if there's no image for the current
game year, and we're refreshing, do an immediate redirect.
Add three new URLS, which redirect in a cycle, one for game SC chart,
one for game SC ownership, and one for game image.