Issue #44: Don't set Game.is_finished from CentreCount.save()
Added a test for issue #44.
Removed CentreCount.save() - we need a full set of CentreCounts before
we can calculate scores for a Game, and Game.save() will recalculate the
scores if is_finished is now True.
Added TODOs to add tests for all the other save() methods, and for where
we may be doing redundant stuff in tests.
In the views, set Game.is_finished (and call Game.save()) when we either
have added CentreCounts for the final game year or have added CentreCounts
including a solo victor.
Also, add an exception handler to Game.news() because when after making the
changes above, some tests were failing because Game.is_finished is now
False when it was previously True.