Add unranked flag to TournamentPlayer
Players with the flag set will not be added to any list returned
by Tournament.best_countries() and they will get the special rank
of UNRANKED in Tournament.scores_and_positions().
In the standings view, they will be listed at the bottom, separate
from the rest of the players.
In the WDD CSV file, they get a special value in the rank column.
Fairly extensive rework to the test code so that I could add tests
of unranked players.