Add support for board seeding
New game_seeder package, modified somewhat from https://github.com/UEWBot/game_seeder.
Add a "seed_games" flag to Tournament, to indicate whether players
are to be assigned to games by the system.
Add a "game_count" to RoundPlayer to track which players are going
to sit out the round or to play multiple games.
Two new template pages, to specify who will sit out or play two games,
and to show seeded games and allow powers to be assigned to the powers.
Two corresponding new views, and some forms.

This completes the "frist version" fix for bug #35.