Models: Add Round.in_progress()
This helps when there could be earlier rounds ongoing when a new
round starts. A round becomes "in progress" as soon as RoundPlayers
are created (usually at roll call) or Games are created. It remains
"in progress" until is_finished() returns True (i.e. it has Games and
they've all finished).
Re-work Tournament.current_round() to prioritise the last round that
is still in progress over the first round that is not yet finished.

Some major rework was needed in the test code because the current round
for t1 is now round 2, not round 1 (which seems more reasonable).