Add the ability to bias the game seeding
Fixes bug #51

Add a new model, SeederBias, that pairs two TournamentPlayers
with a weight. Add the new model to the Admin.
Add a new method, GameSeeder.add_bias() to add biasing to the
seeding algorithm. The two players are treated as if they've
already played the specified number of games together, and so
will be kept apart when seeding boards.
In _seed_games(), add all biases for the tournament to the seeder
before doing the seeding.