Models: Rework Tournament scoring methods
In the tournament_scores() view, we need the overall current
tournament score and the score for each round (including any
currently in progress). When the round is over these are just
retrieved from the database, but when it's ongoing they are
calculated from the current game positions. Calculating the
current tournament score also involves retrieving all the round
scores, including recalculating the score for any ongoing rounds.
The net effect is that we calculate the score for every ongoing
game and round twice (once for the round scores and again for the
tournament scores).
To avoid this, we need to pass the round scores back along with
the tournament scores we calculated from them. This is a fairly
extensive change:
- TournamentScoringSystem.scores() is reworked to return the round
  scores along with the tournament scores. For clarity, it is also
  renamed from score() to scores_detail().
- Tournament.scores() is split into three - calculated_scores()
  will always recalculate and just returns the tournament scores.
  It is called where scores(True) used to be and doesn't care
  whether the tournament is over or not. In practice, it's called
  from Game.save() after the last game has been set to finished and the
  score recalculated and then the last round has been set to finished
  and the score recalculated, so it will always use the stored
  round scores.
  scores_detail() has the same functionality as the old
  scores(force_recalculation=False), but now returns both the
  tournament scores and the round scores.
  _scores_detail_calculated() is the code that is common to both those
  methods.
  The old Tournament.scores() method is removed for clarity.
- Tournament.positions_and_scores() now also returns the round scores.
- all callers of scores() and positions_and_scores() are modified
  accordingly. This mostly affects tournament_scores(), which no
  longer needs to call Round.scores().