Game_Seeder: Add support for powers played
In the constructor, take a list of powers.
In add_played_game(), the game is now a set of (player, power)
    2-tuples.
New method _power_fitness() scores a game (represented as (player,
    power) 2-tuples) based on the number of players who have already
    played those powers.
New exception PowersNotUnique that can be raised by add_played_game().
views and test_game_seeder are migrated to the new interface, and tests
are added for the new code.

Note that code to assign powers to players is not yet added.

Partially addresses Bug #50.