Allow players to enter/update their power prefs
Add a uuid_str attribute to TournamentPlayer.
Add new URLs to enter preferences for a single player, using the
uuid_str as part of the URL for secrecy.
Add a send_prefs_email() function to send the URL to the player.
(could be a TournamentPlayer method, but I like keeping all the
email-related code together).
Call that new function when a new TournamentPlayer is first saved.
(We'll probably want to send it in other places, too)
Add a new form for one player to enter their preferences.
Add a new view function, player_prefs().

Partly addresses Bug #49.

Squash