Views: Clean up Preferences forms
Both PrefsForm and BasePrefsFormset now provide an implicit
'initial' derived from the TournamentPlayer or Tournament
respectively. Both also support overriding that inplicit
value if needed.
Views are changed to use the default initial value.

Also, change the views to check for changes and to only
update preferences that are changed in the form (if any).