Forms: Improve GamePlayerForm error message and add tests
Player.__str__() is much more meaningful than RoundPlayer.__str__()
in this context (we're picking from RoundPlayers, so the non-Player
part is the same for all choices), so display that.
Rename player to r_player and players to r_players for clarity
(they are RoundPlayers, not Players).

Add tests for GamePlayerForm.