Forms: Fix, rename, and test BaseGamePlayersForm
Rename to BaseGamePlayersFormset, because it is a FormSet class, not
a Form class.
Re-write the clean() method so that it does what it should (In practice,
because we set the number of forms on the page in the view, this code
only really gets called when the formset is valid and fully populated,
so this code was untested).
Change calls to super() to not pass parameters.
Add tests of the class.