BasePowerAssignFormset tests and fixes
Rename BasePowerAssignForm to BasePowerAssignFormset because it's
a FormSet, not a Form.
Add assertions to __init__() because the formset is only useful
for assigning GreatPowers to GamePlayers of an existing Game, so
not setting extra to zero, or passing a Round with no Games are
both erroneous uses.
Simplify the clean() method.
Add test cases.

Also remove a test constant that was only used in one place.