Views: Redirect from seed_games to get_seven if necessary
For a GET of seed_games, if we do not have an exact multiple of
7 players, allowing for players sitting out or playing two games,
redirect to the get_seven page.
In the roll_call page, don't repeat this check, just go straight
to the seed_games page and allow it to redirect if necessary.
Note that we leave the check in get_seven() alone, because we
may open that view to change which players are sitting out or
playing two games.
Remove the link to get_seven from the round detail page - the
seed_games link will no wtake us there if needed.
Add a test of seed_games with an odd number of players.

Fixes bug #81.