Player: Cache WDD name
The WDD name doesn't change very often, and retrieving it from
the WDD can take a long time, which makes the player index page,
in particular, very slow. So add a cache of the WDD name to the
Player class, and populate it on-demand. Also update it from
Player.save() in case the wdd_id has been changed.