Views: Minor clean-up
In the player_detail view, add a comment that we're not bothering
to check permissions.
In the tournament player view, add a check that the Tournament is
still running and editable to the "POST" path.
In the tournament player template, add links to the player details
to the read-only version. It probably makes sense to add links to
the modifiable version, too, but I haven't decided whether to link
to the admin page for the player or elsewhere, and there are still
other changes coming there anyway.
In three templates, fix the classes of table rows.