Views: Add dropdowns to register players
New PlayerForm form.
Add formset display to the template.
Add code to create the formset to the view (do this unconditionally,
because it's cheap, even though it won't be used if the user can't
edit the Tournament), and process any posted data.
The only possible error is if you try to register a player who's
already registered. In this case, taht entry will just be ignored,
which seems ok.

Missing test code.