Models: Fix Game.save() where the_set is changed
As the code stands, we try to retrieve the image for the initial
position of the new set, which doesn't exist, so we get a new
GameImage and the save() of that fails the uniqueness constraint.
We need to retrieve the GameImage for the initial position (if any),
then set the image, and save it.