Views: Fix power assignment in seed_games()
If the user changes the powers assigned to players, we'd
get a uniqueness constraint violation because it updates the
GamePlayers one at a time, so there woudl briefly be two players
for the same power. Fix that by unassining all powers before
processing the form fields.
Also add several unit tests for POSTs in get_seven() and
seed_games().