Views: Fix roll_call for specific round
The test for "is it the current round?" was broken, always
evaluating to False.
For the roll cal of a completed round, we need a redirect to
a GET. The only reasonable place to redirect is the same page.
Add a TODO comment - should we reset each RoundPlayer's game_count
to 1? If we're changing the roll call, we may well want to re-do
the determination of who is sitting out or playing two games, too...
(particularly in the case where somebody has agreed to play two games
and then a new player shows up). Of course if we've already done the
"get seven" and then the roll call changes, presumably we now don't
have an exact multiple of 7, so we'll have to re-do it anyway, and
knowing who had agreed to sit out or play two games is valuable. If
we haven't done the "get seven", all game_counts will still be 1 anyway.
Also add several roll_call POST tests.