Add user attribute to Player class
This allows us to track which Players are also users on the site,
and thereby to figure out that a TournamentPlayer is also a manager
for a Tournament.
Automatically set the unranked attribute when adding a manager as
a TournamentPlayer.
List whether TournamentPlayers are unranked in the tournament players
view. Also add some missing whitespace to the template.

Helps address Bug #92.