Views: Add elimination year to SC count form
Some TDs want to just enter the final SC counts for each board.
That works fine unless they're using a scoring system where
elimination order or elimination year matters. For those cases,
add an (optional) "elimination year" field to each column. If a year
is entered, create a SupplyCentreCount for that power and year with
a count of zero.

Fixes Bug #91