Tests: Fix issues with DrawProposals in models test
Ensure that DrawProposals are only added to the database when necessary.
Ensure that they are deleted at the end of the test when they are added
to the database by a test.
In test_draw_proposal_multiple_successful(), give the 2 proposals
different years.
Adding a DrawProposal to the database can change the state of the Game.
Undo any such changes at the end of the test.