Add self-check-in capability
Still some work to do.
Add send_roll_call_email() to send individualised emails with the
TournamentPlayer.get_prefs_url().
Add EnableCheckInForm to provide a UI to enable (or disable)
self-check-ins.
Add enable_check_in and email_sent BooleanFields to Round class.
Add a "round__start"" ordering to RoundPlayers, so they are
ordered sensibly for a single Player.
In roll_call(), also create an EnableCheckInForm and pass it to the
template. Set enable_check_in for the Round as appropriate. The first
time it is set True, also call send_roll_call_email() for each
TournamentPlayer and set email_sent to True.
Rename player_prefs.heml to player_entry.html. Make the preferences
part conditional on t.power_assigned_from_prefs(). Add a second
section that lists each round. Note whether the player is registered
for the round or not. Also check whether enable_check_in is set if
the round is not finished. This is where we need to add a button to
do the self-check-in.
In roll_call.html, also display the (enable check in) form.
In player_prefs, also create a list of (Round, RoundPlayer) 2-tuples
and pass it to the template. We also need to handle self-check-ins here.