Models: Replace discord username with backstabbr username
Discord usernames aren't fixed, so there's little point in storing
them at all. For a VFTF tournament, though, backstabbr usernames
are useful for mapping players on the two systems.
They could also be useful if we ever want to pull in online data
to add to the player background stuff.
This change just replaces the discord usernames with backstabbr
usernames.