Game Scoring: Add ManorCon system
As described by Don Del Grande:
Solo - 75 points; all others get 0.1 per season survived
Eliminated players get 0.1 per season survived
Each player in a draw divides 100 points in proportion to (N squared + 4N + 16), where N is their SC count
This is the tricky part: each eliminated player counts as 16 towards the total, even though they don't get any points other than their survival points.
For example, let's say it's a 12-10-8-4 draw.
The 12 is worth 144 + 48 + 16 = 208 toward the total
The 10 is worth 100 + 40 + 16 = 156 toward the total
The 8 is worth 64 + 32 + 16 = 112 toward the total
The 4 is worth 16 + 16 + 16 = 48 toward the total
The three eliminations are each worth 16 toward the total
The sum of these is 208 + 156 + 112 + 48 + (3 x 16) = 572
The tournament points are:
For the 12: 100 x 208 / 572 = 36.3636
For the 10: 100 x 156 / 572 = 27.2727
For the 8: 100 x 112 / 572 = 19.5804
For the 4: 100 x 48 / 572 = 8.3916
Note that the numbers don't add up to 100. They would, if each elimination counted as 100 x 16 / 572 = 2.7972.

Issue #143