Game views: Add new scrape_backstabber()
New view function scrape_backstabbr() that will grab the current
centre counts if there's a backstabbr URL in the game's notes
attribute and create/update the appropriate CentreCount objects.
Add the corresponding new URL and template (that lists the new CentreCounts).
Link to the new page from the game details page if the tournament
is virtual and the game has notes.

Helps with issue #140.