Views: Change default game names
For vWDC, they want to use letters to identify games rather than numbers.
Ideally, this would be controlled by a new Tournament flag, but keep it
simple for now. The plan is to revert this after vWDC is over, and then
to do it in a better way later on.

Issue #159.