Views: Fix number of blank rows in game forms
Both the form to enter the supply centre counts for a game
and the form to enter supply centre ownerships for a game
should provide a reasonable number of blank rows.
If the game ends at a fixed point, we should provide one row
for each possible game year that hasn't already been entered.
Also, we can share this code between the two views.

Issue #167.