Views: Reset RoundPlayer.game_count in roll_call()
If we've returned to the roll call page, we may have previously set
game_count for some RoundPlayers. We need to clear that (a) because
the number of players may have changed, which means that we need
different numbers of players to sit out or play two games and (b)
because get_seven() ignores game_count while seed_games() doesn't,
leading to a possible redirect loop.
The expectation here is that game_count is effectively a way for
get_seven() to communicate with seed_games().

Issue #177