Views: Speed up simple TournamentPlayer index page
Rather than creating a formset only to ignore it when rendering
the template, check permissions in the view code and have two
distinct templates. This should be far faster.
There is one minor change in behaviour - a POST to that page by
a user without permission will now get the read-only page (same
as for a GET) rather than a 404 error.