Models: Remove replacement player support
This was never fully implemented (indeed, a lot of the removals
are comments saying that code doesn't work if there are replacement
players).
The Player playing a GreatPower in a Game can still be changed, but
only a complete replacement, as if the original Player was never
playing that power.

Issue #4