Remove faulty error-handling
The add_error() call parameters are clearly wrong.
TournamentPlayer has no clean() method, and the only uniqueness
constraint is already satisfied in these cases. That means that
full_clean() could only generate errors as a result of field
validation. I tried and failed to get that to happen, so it looks
like these error conditions can never be reached.

RoundPlayer.clean() just checks for an existing TournamentPlayer,
which in this case we know exists. There are no uniqueness constraints
on RoundPlayer.