Error handling in game_scores() view
If we take a valid GamePlayer and update just the score attribute,
there's no way for it to them be invalid.
Also, by using filter() and update() we can affect all players of
the given power if there were replacements (they all get the same
score, but how scoring works for games with replacement players
is unspecified anyway, and at least we avoid the exception we would
get using get()).