Remove useless error checking in roll_call view
RoundPlayer.clean() just checks that a corresponding
TournamentPlayer exists, which is always guaranteed in this
code block. It has no uniqueness constraint (although perhaps it
should - for the player/round combo, but even then we're using
get_or_create() so we're guaranteed not to violate it). Form does
field validation.