Remove unnecessary error handling in round views
GamePlayer has no uniqueness constraints (there probably should
be one for the player/game pairing - see Issue #178). Clean()
is mostly concerned with replacement players, which doesn't work.
The remaining part of clean() is to check that there is already
a corresponding TournamentPlayer. Given that in one case we're using a
RoundPlayer as the field value, and RoundPlayer checks that there's
a corresponding TournamentPlayer, and in the other we're changing an
existing GamePlayer, this will always be ok.

Fixup round_views error handling