Fix standby players selection
Distinguish between the case where we need all standby players
to play and the one where we need some of them (no need to pick
them in the former case).
Don't duplicate code between the form and the view - use the form
calculations in the view.
Clean up the wording on the page - always state the number wanting
to play and number of standbys and then follow with the conclusions.

Issue #169