Backstabbr: Replace Game.powers with sc_counts and players
Rather than the (dots, player, profile URL) powers 3-tuple,
store a (player, profile URL) players 2-tuple and a separate sc_counts.
Both are dicts indexed by power.
This simplifies the code to construct them and the code that uses
them, and it separates the part that varies through the game from
the part that is fixed.