Initial work to display issues with board seeding
When seeding games and powers, also return a list of issues for
each game. For now, the list is either empty or contains a single
entry saying how many players have already played their assigned
power. In future, it should detail who has played which power before.

Display the list of issues in the form.

Issues #220