Views: Change behaviour of roll call page for old rounds
Issue #233.

Now that we always go to the "get seven" page from the roll call page,
there's no need for the check that we're changing the current round.
We already have a warning on the roll call page if games exist and a
warning on the seed games page that games will be deleted. The current
behaviour, where we redirect to the same page (as a GET) with no
explanation is confusing for users.