Players: Streamline Player.save()
Remove the call to add_player_bg() from Player.save(). Players are
saved more frequently than they used to be, and we now have a button
to update the background on demand, so we can live without this call.
Add a cache of wdd_player_id so we can check whether it has changed.
Don't re-populate _wdd_name, just clear it if it's now wrong. It will
get populated in wdd_name() when needed.

Issue #243