Models: Store "if all games ended now" scores in the database
Issue #234

Rather than only storing a score in the database when a Game (or Round
or Tournament) ends, store it whenever a Game progresses. This decreases
the number of times we have to do the calculation, and opens up
possibilities for things like using font to show which are final or
interim scores.
This first patch just changes the models and game views to do the
storage and retrieval correctly.