Views: Convert URLs from re_path() to path()
Add two URL converters, for game years and game turns.

This changes one URL from valid to 404:
  tournaments/<id>/games/<name>/timelapse//
but it was a bug that that URL was valid anyway, and images
aren't actually being used in practice to my knowledge.
  tournaments/<id>/games/<name>/timelapse/
remains valid.
Had to tweak one one game view because the reverse mapping changed.

Issue #82