Add Circuits support
Add models to support circuits.
A Circuit is a met event made up of a number of Tournaments. Tournament
scores are combined to create a circuit score for each player.
A CircuitSeries is a series of Circuits, generally where a Circuit is
repeated over a number of years.
A CircuitPlayer is a Player who participated in a Circuit.

Issue #281