Models: Move code from save() to field pre_save() methods
In Series, TournamentPlayer, and DrawProposal, remove code that changes
field value from the custom save() method and instead use a custom
pre_save() method for the corresponding field.

This avoids the need to amend update_fields in the save() method.

Note that we also remove the raising of ValidationError from
DrawProposal.save(). The check is no longer reliable because is_passed may
not be set until the superclass's save() method calls the field's pre_save(),
which happens after this check. I don't want to duplicate the
code that we moved to pre_save(). In practice, the database constraint
will prevent the actual save, we'll just get an IntegrityError instead.
Therefore that test case is changed to check DrawProposal.full_clean()
rather than save(). The code calling save() always calls full_clean() first.

Issue #277