Add support for no round scoring system
This is useful when using a tournament scoring system that looks at
games rather than rounds. In those cases, when calculating the
tournament scores, we overwrite any previously-calculated round scores
anyway, so it's much better to not have a round scoring system at all.

So this adds the option of "None" as a round scoring system, with the
corresponding object also being None. In that case it won't be displayed
on the tournament details page and round scores will not be calculated
(but tournament scores still will be calculated).

Issue #139