Views: Display scores for some tournaments differently
If a tournament scoring system considers all games equally regardless
of the round in which they were played, display game scores in
preference to round scores. If a player didn't play in a particular
round but was present (i.e. a RoundPlayer exists but no GamePlayer),
display the RoundPlayer.score because they may have gotten points for
volunteering to sit out the round.

There's a lot of redundancy between the new template and the existing
one, but that can be addressed later.

Issue #316