Background: Add support for reading the WDR
Similar to reading the WDD, we can provide background on the players.

Note that some Tournaments have different names in the WDD and WDR, and
the system cannot currently identify these as being the same tournament,
so you may end up with duplicate PlayerTournamentRankings and
PlayerGameResults if you add a wdr_player_id to a Player who already
has a wdd_player_id and who has already had background added and then
you update their background. It may make sense to remove all
PlayerTournamentRankings and PlayerGameResults for the Player
beforehand in this case.