Background: Separate titles from tournament rankings
Create a new PlayerTitle class. Migrate titles from
PlayerTournamentRankings to the new class.
This avoids issues when the system doesn't match up a title (from
wikipedia) with the tournament (from the WDD/WDR), which results
in the tournament being counted multiple times. It also better
handles cases where DipCon and WDC titles were awarded together.

The new class has an optional ForeignKey to PlayerTournamentRanking,
which is not populated by the system at this time.