Views: Send fewer emails when Games are changed
Distinguish between changes to Games and GamePlayers in the two
forms to modify games. This allows us to only update the necessary
models, and also to only re-send the board call email if the
players or power assignments have changed, but not if things like
the external URL have changed.

To use form.changed_data, we need to ensure that the initial parameter
passed to the formset constructor is correct.

Issue #349