Backstabbr: Correctly parse games that end in winter
When we ask for a specific turn, backstabbr will redirect to the
default page for the game if the turn we request is the current
turn. So retrieving the winter turn of a game that ended in that
winter phase currently raises InvalidGameUrl.

By following redirects, we can parse the turn correctly.

Backstabbr will return 404 errors for malformed URLs, so this doesn't
cost us any functionality.