Check permissions for individual tournaments
In general, only users who are managers for a given tournament can
modify or delete it. In addition, don't allow tournaments with
is_editable = False to be edited or deleted and don't allow
tournaments with games to be deleted. These checks also apply to
the various other objects within the tournament.

Issue #8
Issue #181