DrawProposal-related fixes
In DrawProposal, always use self.game.survivors(self.year-1) to get
the list of surviving powers. Note that a draw in a given year will
use the status at the end of the previous year.
In news.py, use ngettext() to deal with draws vs concessions.
Improve tests of DrawProposal.votes_against().
Add a new test of draw proposals in _game_news().