Add graph of tournament score over time (rounds)
Adds two new view functions (one for the graph itself and one
for the page containing it) and corresponding template and URLs.
Adds links to the score graph from the views page and the
tournament details page.

Would be nice to compress the graph itself horizontally -
most tournaments are only 3 or 4 rounds.

Issue #249