Add support for optional Pools within Rounds
Add a new Pool class that is associated with a Round.
A Round will usually have no Pools but may optionally have two,
one with and one without the board_count attribute set.
After roll call, if the round has pools, the TD will be asked to
assign players to the pool with the board_count attribute.
"get_seven" is restricted to pools without the board_count attribute
when the round has pools.
Seeding looks the same from the TD's perspective, but behind the
scenes is done for each pool separately, completely independent of
one another.